refactor(bundler): move type bundler to a separate file
